pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tic_tac_toe_web_interface"
version = "0.1.0"
description = "Tic Tac Toe Web Interface"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
    "numpy>=2.2.0",
    "msal>=1.32.0",
    "PyJWT>=2.10.0",
    "cryptography>=45.0.0",
    "tic_tac_toe_game @ git+https://github.com/andrewbeaton123/tic_tac_toe_game.git"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.15.0",
    "ruff>=0.5.0",
    "pre-commit>=3.5.0",
    "types-requests"
]

[tool.hatch.metadata]
allow-direct-references = true
